Transition to QMake/Make
Rebuilt the project in QT to confirm it executes. With the exclusion of Course_Scheduler.pro.user
