+++ course.h
@@ -1,0 +1,4 @@
+#ifndef COURSE_H
+#define COURSE_H
+
+#endif // COURSE_H
