+++ department.h
@@ -1,0 +1,4 @@
+#ifndef DEPARTMENT_H
+#define DEPARTMENT_H
+
+#endif // DEPARTMENT_H
