+++ instructor.h
@@ -1,0 +1,4 @@
+#ifndef INSTRUCTOR_H
+#define INSTRUCTOR_H
+
+#endif // INSTRUCTOR_H
