+++ resourceManager.h
@@ -1,0 +1,4 @@
+#ifndef RESOURCEMANAGER_H
+#define RESOURCEMANAGER_H
+
+#endif // RESOURCEMANAGER_H
