+++ room.h
@@ -1,0 +1,4 @@
+#ifndef ROOM_H
+#define ROOM_H
+
+#endif // ROOM_H
