--- source/course.cpp
@@ -1,1 +1,0 @@
-#include "../headers/course.h"
