--- source/department.cpp
@@ -1,1 +1,0 @@
-#include "../headers/department.h"
