--- source/instructor.cpp
@@ -1,1 +1,0 @@
-#include "../headers/instructor.h"
