--- source/resourceManager.cpp
@@ -1,10 +1,0 @@
-#include "../headers/course.h"
-#include "../headers/department.h"
-#include "../headers/engine.h"
-#include "../headers/instructor.h"
-#include "../headers/room.h"
-
-int main()
-{
-
-}
